refactor: dropdown examples
