+++ src/showcase/app/public/public-examples/dropdown-examples/dropdown-custom-trigger-example/dropdown-custom-trigger-example.component.ts
@@ -1,0 +1,30 @@
+import { Component } from '@angular/core';
+
+interface DropdownLink {
+  page: number;
+  text: string;
+}
+
+@Component({
+  selector: 'sbb-dropdown-custom-trigger-example',
+  templateUrl: './dropdown-custom-trigger-example.component.html',
+  styleUrls: ['./dropdown-custom-trigger-example.component.css'],
+})
+export class DropdownCustomTriggerExampleComponent {
+  links: Array<DropdownLink> = [
+    { page: 1, text: 'Test 1' },
+    { page: 2, text: 'Test 2' },
+    { page: 3, text: 'Test 3' },
+  ];
+
+  linkGenerator(page: number) {
+    return {
+      queryParams: { page: page },
+      routerLink: ['.'],
+    };
+  }
+
+  onClick() {
+    console.log('Button clicked');
+  }
+}
