--- src/showcase/app/public/public-examples/dropdown-examples/dropdown-example/dropdown-example.component.html
@@ -1,120 +1,0 @@
-<div class="container-fluid">
-  <div class="row">
-    <div class="col-lg-6 sbbsc-v-sep">
-      <h4 class="sbbsc-block">Button trigger</h4>
-      <button sbbButton type="button" [sbbDropdown]="dropdown">
-        Open dropdown
-      </button>
-      <sbb-dropdown #dropdown="sbbDropdown">
-        <a
-          *ngFor="let link of links"
-          sbbDropdownItem
-          [routerLink]="linkGenerator(link.page).routerLink"
-          [queryParams]="linkGenerator(link.page).queryParams"
-          routerLinkActive="sbb-selected"
-        >
-          {{ link.text }}
-        </a>
-
-        <hr />
-        <button sbbDropdownItem type="button" (click)="onClick()">
-          Abmeldung
-        </button>
-      </sbb-dropdown>
-    </div>
-    <div class="col-lg-6">
-      <h4 class="sbbsc-block">Properties</h4>
-      <div class="row">
-        <div class="col-12 sbbsc-block">
-          <button type="button" sbbButton (click)="addNewLink()">
-            Add new link
-          </button>
-        </div>
-      </div>
-    </div>
-  </div>
-
-  <div class="row">
-    <div class="col-lg-6 sbbsc-v-sep">
-      <h4 class="sbbsc-block">Custom trigger</h4>
-      <div
-        class="dropdown-avatar"
-        [sbbDropdown]="dropdown2"
-        tabindex="0"
-        [class.dropdown-avatar-active]="dropdown2.open"
-      >
-        <div class="avatar-icon">
-          <sbb-icon-user></sbb-icon-user>
-        </div>
-        <div *ngIf="dropdown2.open">William Tell</div>
-      </div>
-      <sbb-dropdown #dropdown2="sbbDropdown">
-        <a
-          *ngFor="let link of links"
-          sbbDropdownItem
-          [routerLink]="linkGenerator(link.page).routerLink"
-          [queryParams]="linkGenerator(link.page).queryParams"
-          routerLinkActive="sbb-selected"
-        >
-          {{ link.text }}
-        </a>
-        <hr />
-        <button sbbDropdownItem type="button" (click)="onClick()">
-          Abmeldung
-        </button>
-      </sbb-dropdown>
-    </div>
-  </div>
-
-  <div role="presentation" class="sbbsc-separator"></div>
-
-  <div class="row">
-    <div class="col-lg-6 sbbsc-v-sep">
-      <h4 class="sbbsc-block">Dropdown opened from elsewhere</h4>
-
-      <div
-        sbbDropdownOrigin
-        #dropdownOrigin="sbbDropdownOrigin"
-        class="dropdown-avatar"
-        tabindex="0"
-        [class.dropdown-avatar-active]="dropdown3.open"
-      >
-        <div class="avatar-icon">
-          <sbb-icon-user></sbb-icon-user>
-        </div>
-        <div>William Tell</div>
-      </div>
-      <sbb-dropdown #dropdown3="sbbDropdown">
-        <a
-          *ngFor="let link of links"
-          sbbDropdownItem
-          [routerLink]="linkGenerator(link.page).routerLink"
-          [queryParams]="linkGenerator(link.page).queryParams"
-          routerLinkActive="sbb-selected"
-        >
-          {{ link.text }}
-        </a>
-        <hr />
-        <button sbbDropdownItem type="button" (click)="onClick()">
-          Abmeldung
-        </button>
-      </sbb-dropdown>
-    </div>
-    <div class="col-lg-6">
-      <h4 class="sbbsc-block">Separated trigger from dropdown origin</h4>
-      <div class="row">
-        <div class="col-12 sbbsc-block">
-          <button
-            mode="ghost"
-            type="button"
-            [sbbDropdownConnectedTo]="dropdownOrigin"
-            [sbbDropdown]="dropdown3"
-            sbbButton
-          >
-            Open dropdown from here
-          </button>
-        </div>
-      </div>
-    </div>
-  </div>
-</div>
