--- src/showcase/app/public/public-examples/dropdown-examples/dropdown-example/dropdown-example.component.scss
@@ -1,27 +1,0 @@
-$borderActive: solid 2px #666;
-$border: solid 2px white;
-$sbbColorGrey: #666;
-.dropdown-avatar {
-  border: $border;
-  display: flex;
-  align-items: center;
-  padding: 14px;
-  color: $sbbColorGrey;
-
-  svg > path {
-    stroke: $sbbColorGrey;
-  }
-
-  .avatar-icon {
-    width: 24px;
-    height: 24px;
-  }
-  &-active {
-    border-top: $borderActive;
-    border-right: $borderActive;
-    border-left: $borderActive;
-    &.sbb-dropdown-trigger-above {
-      border-bottom: $borderActive;
-    }
-  }
-}
